Add multiple config file support addressing #2
The configuration files are considered from first to last (order in the
argument list). New keys are added and existing keys are merged.

Currently you CAN NOT add to an array, it will overwrite it.
i.e. if you want to update the "memory" field, you have to specify all
the fields again. Might need future improvements, but this is ok for
"plugin specific" configuration files