[Plugin][IdempotencyStatistics] Use new cycle count feature
Replace all the instruction count values with cycle count.
This can be disabled by setting `use_cycle_count` to false.

Although cycle count should be a better way of benchmarking, it wont
affect the war results at all. But it might give a better idea of the
overhead AND how far apart the WAR violations are (more interesting).